Make RemoteStore not extend Store
This allows us to get rid of the then unused Store class.
