--- k9mail-library/src/main/java/com/fsck/k9/mail/Store.java
@@ -1,51 +1,0 @@
-
-package com.fsck.k9.mail;
-
-
-import java.util.List;
-
-/**
- * Store is the access point for an email message store. It's location can be
- * local or remote and no specific protocol is defined. Store is intended to
- * loosely model in combination the JavaMail classes javax.mail.Store and
- * javax.mail.Folder along with some additional functionality to improve
- * performance on mobile devices. Implementations of this class should focus on
- * making as few network connections as possible.
- */
-public abstract class Store {
-    public abstract Folder<? extends Message> getFolder(String name);
-
-    public abstract List <? extends Folder > getPersonalNamespaces(boolean forceListAll) throws MessagingException;
-
-    public abstract void checkSettings() throws MessagingException;
-
-    public boolean isCopyCapable() {
-        return false;
-    }
-
-    public boolean isMoveCapable() {
-        return false;
-    }
-
-    public boolean isPushCapable() {
-        return false;
-    }
-
-    public boolean isSendCapable() {
-        return false;
-    }
-
-    public boolean isExpungeCapable() {
-        return false;
-    }
-
-    public boolean isSeenFlagSupported() {
-        return true;
-    }
-
-    public void sendMessages(List<? extends Message> messages) throws MessagingException { }
-
-    public Pusher getPusher(PushReceiver receiver) {
-        return null;
-    }
-}
